Move utils into folder
