+++ components/map.js
@@ -6,7 +6,7 @@ import { useThemedColormap } from '@carbonplan/colormaps'
 import { Minimap, Path, Sphere } from './minimap'
 import { PROJECTIONS, ASPECTS } from './constants'
 import useStore from './data/store'
-import { getMapProps, getProjection, validatePoint } from './utils'
+import { getMapProps, getProjection, validatePoint } from './utils/data'
 import MapContainer from './map-container'
 import Layer from './minimap/layer'
 import Nav from './nav'
