+++ components/region/point.js
@@ -1,7 +1,7 @@
 import { useThemeUI } from 'theme-ui'
 import { useCallback, useEffect, useMemo, useRef } from 'react'
 import useStore from '../data/store'
-import { getProjection } from '../utils'
+import { getProjection } from '../utils/data'
 import { ASPECTS } from '../constants'
 
 const Point = ({ mapProps }) => {
